feat: renamed crm to procurement
The CRM module was a bit confusing, as it was meant to cover the procurement process.
New names are:
- CRM -> PROCUREMENT
- LEADS -> REQUIREMENTS
- OPPORTUNITIES -> EVALUATIONS

I think this better reflects the purchase process of assets and licenses. Also, note that the internal paths are still referencing the old terminology and requires a refactor, which will be done in the next couple of prs